(Sem Teste) Adicionado logs e relacionamentos
