+++ App/Models/Login.cs
@@ -7,4 +7,6 @@ public class Login
     public string Senha { get; set; } = string.Empty;
     public string? CodigoAutenticacao { get; set; }
     public int IdUsuario { get; set; }
+
+    public Usuario Usuario { get; set; }
 }
