+++ App/Models/Sexo.cs
@@ -4,4 +4,6 @@ public class Sexo
 {
     public int Id { get; set; }
     public string Descricao { get; set; }
+    
+    public List<Usuario> Usuarios { get; set; }
 }
